printf finished

+++ ft_puthex_min.c
@@ -6,13 +6,13 @@
 /*   By: juanrome <juanrome@student.42madrid.fr>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/06 14:27:17 by juanrome          #+#    #+#             */
-/*   Updated: 2025/01/08 21:51:03 by juanrome         ###   ########.fr       */
+/*   Updated: 2025/01/09 19:39:45 by juanrome         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int	ft_puthex_min(unsigned int nb)
+int	ft_puthex_min(unsigned long long nb)
 {
 	int		re;
 	char	*base;
